refactor: move GalleryImage interface to types file
Moves GalleryImage interface from utils/gallery.ts to types/gallery.ts
to centralize all gallery-related type definitions.

Changes:
- Add GalleryImage interface to src/types/gallery.ts
- Import GalleryImage from types in utils/gallery.ts
- Update PhotoSwipeGallery.astro to import from types
- Add ImageMetadata import to types file

Benefits:
- All gallery types now centralized in one location
- Better separation: types in types/, logic in utils/
- Consistent with other gallery types (LayoutBox, etc.)
- More discoverable for developers (all types in one place)
