+++ src/components/PhotoSwipeGallery.astro
@@ -1,6 +1,6 @@
 ---
 import { Image, getImage } from "astro:assets";
-import type { GalleryImage } from "@utils/gallery";
+import type { GalleryImage } from "@mttypes/gallery";
 import 'photoswipe/style.css';
 import '@css/photoswipe.css';
